Fix ICC build break.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@8158 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Universal/SetupBrowserDxe/Setup.c
@@ -2225,7 +2225,7 @@ GetIfrBinaryData (
             // Check ClassGuid of formset OpCode
             //
             IsSetupClassGuid  = FALSE;
-            NumberOfClassGuid = ((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3;
+            NumberOfClassGuid = (UINT8) (((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3);
             ClassGuid         = (EFI_GUID *) (OpCodeData + sizeof (EFI_IFR_FORM_SET));
             for (Index = 0; Index < NumberOfClassGuid; Index++) {
               if (CompareGuid (ClassGuid + Index, &gEfiHiiPlatformSetupFormsetGuid)) {
